requirements.txt
Django>=3.2.1,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
Celery 
redis
